seperate the vm_manager into several slice files.
