+++ src/bosh_azure_cpi/lib/cloud/azure.rb
@@ -64,6 +64,8 @@ require 'cloud/azure/network/manual_network'
 require 'cloud/azure/network/dynamic_network'
 require 'cloud/azure/network/vip_network'
 require 'cloud/azure/network/network_configurator'
+require 'cloud/azure/vms/vm_manager_availability_set'
+require 'cloud/azure/vms/vm_manager_network'
 require 'cloud/azure/vms/vm_manager'
 require 'cloud/azure/storage/blob_manager'
 require 'cloud/azure/disk/disk_manager'
