+++ src/bosh_azure_cpi/lib/cloud/azure/vms/vm_manager.rb
@@ -425,278 +425,10 @@ module Bosh::AzureCloud
       Base64.strict_encode64(JSON.dump(user_data))
     end
 
-    def _get_network_subnet(network)
-      subnet = @azure_client.get_network_subnet_by_name(network.resource_group_name, network.virtual_network_name, network.subnet_name)
-      cloud_error("Cannot find the subnet '#{network.virtual_network_name}/#{network.subnet_name}' in the resource group '#{network.resource_group_name}'") if subnet.nil?
-      subnet
-    end
-
-    def _get_network_security_group(vm_props, network)
-      network_security_group = nil
-      # Network security group name can be specified in vm_types or vm_extensions, networks and global configuration (ordered by priority)
-      network_security_group_name = vm_props.security_group.nil? ? network.security_group : vm_props.security_group
-      network_security_group_name = @azure_config.default_security_group if network_security_group_name.nil?
-      return nil if network_security_group_name.nil?
-      cloud_error('Cannot specify an empty string to the network security group') if network_security_group_name.empty?
-      # The resource group which the NSG belongs to can be specified in networks and global configuration (ordered by priority)
-      resource_group_name = network.resource_group_name
-      network_security_group = @azure_client.get_network_security_group_by_name(resource_group_name, network_security_group_name)
-      # network.resource_group_name may return the default resource group name in global configurations. See network.rb.
-      default_resource_group_name = @azure_config.resource_group_name
-      if network_security_group.nil? && resource_group_name != default_resource_group_name
-        @logger.info("Cannot find the network security group '#{network_security_group_name}' in the resource group '#{resource_group_name}', trying to search it in the default resource group '#{default_resource_group_name}'")
-        network_security_group = @azure_client.get_network_security_group_by_name(default_resource_group_name, network_security_group_name)
-      end
-      cloud_error("Cannot find the network security group '#{network_security_group_name}'") if network_security_group.nil?
-      network_security_group
-    end
-
-    def _get_application_security_groups(vm_props, network)
-      application_security_groups = []
-      # Application security group name can be specified in vm_types or vm_extensions and networks (ordered by priority)
-      application_security_group_names = vm_props.application_security_groups.nil? ? network.application_security_groups : vm_props.application_security_groups
-      application_security_group_names.each do |application_security_group_name|
-        # The resource group which the ASG belongs to can be specified in networks and global configuration (ordered by priority)
-        resource_group_name = network.resource_group_name
-        application_security_group = @azure_client.get_application_security_group_by_name(resource_group_name, application_security_group_name)
-        # network.resource_group_name may return the default resource group name in global configurations. See network.rb.
-        default_resource_group_name = @azure_config.resource_group_name
-        if application_security_group.nil? && resource_group_name != default_resource_group_name
-          @logger.info("Cannot find the application security group '#{application_security_group_name}' in the resource group '#{resource_group_name}', trying to search it in the default resource group '#{default_resource_group_name}'")
-          application_security_group = @azure_client.get_application_security_group_by_name(default_resource_group_name, application_security_group_name)
-        end
-        cloud_error("Cannot find the application security group '#{application_security_group_name}'") if application_security_group.nil?
-        application_security_groups.push(application_security_group)
-      end
-      application_security_groups
-    end
-
-    def _get_ip_forwarding(vm_props, network)
-      ip_forwarding = false
-      # ip_forwarding can be specified in vm_types or vm_extensions and networks (ordered by priority)
-      ip_forwarding = vm_props.ip_forwarding.nil? ? network.ip_forwarding : vm_props.ip_forwarding
-      ip_forwarding
-    end
-
-    def _get_accelerated_networking(vm_props, network)
-      accelerated_networking = false
-      # accelerated_networking can be specified in vm_types or vm_extensions and networks (ordered by priority)
-      accelerated_networking = vm_props.accelerated_networking.nil? ? network.accelerated_networking : vm_props.accelerated_networking
-      accelerated_networking
-    end
-
-    def _get_public_ip(vip_network)
-      public_ip = nil
-      unless vip_network.nil?
-        resource_group_name = vip_network.resource_group_name
-        public_ip = @azure_client.list_public_ips(resource_group_name).find { |ip| ip[:ip_address] == vip_network.public_ip }
-        cloud_error("Cannot find the public IP address '#{vip_network.public_ip}' in the resource group '#{resource_group_name}'") if public_ip.nil?
-      end
-      public_ip
-    end
-
-    def _get_load_balancer(vm_props)
-      load_balancer = nil
-      unless vm_props.load_balancer.nil?
-        load_balancer_name = vm_props.load_balancer
-        load_balancer = @azure_client.get_load_balancer_by_name(load_balancer_name)
-        cloud_error("Cannot find the load balancer '#{load_balancer_name}'") if load_balancer.nil?
-      end
-      load_balancer
-    end
-
-    def _get_application_gateway(vm_props)
-      application_gateway = nil
-      unless vm_props.application_gateway.nil?
-        application_gateway_name = vm_props.application_gateway
-        application_gateway = @azure_client.get_application_gateway_by_name(application_gateway_name)
-        cloud_error("Cannot find the application gateway '#{application_gateway_name}'") if application_gateway.nil?
-      end
-      application_gateway
-    end
-
-    def _get_or_create_public_ip(resource_group_name, vm_name, location, vm_props, network_configurator)
-      public_ip = _get_public_ip(network_configurator.vip_network)
-      if public_ip.nil? && vm_props.assign_dynamic_public_ip == true
-        # create dynamic public ip
-        idle_timeout_in_minutes = @azure_config.pip_idle_timeout_in_minutes
-        validate_idle_timeout(idle_timeout_in_minutes)
-        public_ip_params = {
-          name: vm_name,
-          location: location,
-          is_static: false,
-          idle_timeout_in_minutes: idle_timeout_in_minutes
-        }
-        public_ip_params[:zone] = vm_props.availability_zone.to_s unless vm_props.availability_zone.nil?
-        @azure_client.create_public_ip(resource_group_name, public_ip_params)
-        public_ip = @azure_client.get_public_ip_by_name(resource_group_name, vm_name)
-      end
-      public_ip
-    end
-
-    def _create_network_interfaces(resource_group_name, vm_name, location, vm_props, network_configurator, primary_nic_tags = AZURE_TAGS)
-      network_interfaces = []
-
-      # Tasks to prepare before creating NICs:
-      #   * preapre public ip
-      #   * prepare load balancer
-      #   * prepare application gateway
-      tasks_preparing = []
-
-      tasks_preparing.push(
-        task_get_or_create_public_ip = Concurrent::Future.execute do
-          _get_or_create_public_ip(resource_group_name, vm_name, location, vm_props, network_configurator)
-        end
-      )
-      tasks_preparing.push(
-        task_get_load_balancer = Concurrent::Future.execute do
-          _get_load_balancer(vm_props)
-        end
-      )
-      tasks_preparing.push(
-        task_get_application_gateway = Concurrent::Future.execute do
-          _get_application_gateway(vm_props)
-        end
-      )
-
-      # Calling .wait before .value! to make sure that all tasks are completed.
-      tasks_preparing.map(&:wait)
-
-      public_ip = task_get_or_create_public_ip.value!
-      load_balancer = task_get_load_balancer.value!
-      application_gateway = task_get_application_gateway.value!
-
-      # tasks to create NICs, NICs will be created in different threads
-      tasks_creating = []
-
-      networks = network_configurator.networks
-      networks.each_with_index do |network, index|
-        network_security_group = _get_network_security_group(vm_props, network)
-        application_security_groups = _get_application_security_groups(vm_props, network)
-        ip_forwarding = _get_ip_forwarding(vm_props, network)
-        accelerated_networking = _get_accelerated_networking(vm_props, network)
-        nic_name = "#{vm_name}-#{index}"
-        nic_params = {
-          name: nic_name,
-          location: location,
-          private_ip: network.is_a?(ManualNetwork) ? network.private_ip : nil,
-          network_security_group: network_security_group,
-          application_security_groups: application_security_groups,
-          ipconfig_name: "ipconfig#{index}",
-          enable_ip_forwarding: ip_forwarding,
-          enable_accelerated_networking: accelerated_networking
-        }
-        nic_params[:subnet] = _get_network_subnet(network)
-        if index.zero?
-          nic_params[:public_ip] = public_ip
-          nic_params[:tags] = primary_nic_tags
-          nic_params[:load_balancer] = load_balancer
-          nic_params[:application_gateway] = application_gateway
-        else
-          nic_params[:public_ip] = nil
-          nic_params[:tags] = AZURE_TAGS
-          nic_params[:load_balancer] = nil
-          nic_params[:application_gateway] = nil
-        end
-        tasks_creating.push(
-          Concurrent::Future.execute do
-            @azure_client.create_network_interface(resource_group_name, nic_params)
-            @azure_client.get_network_interface_by_name(resource_group_name, nic_name)
-          end
-        )
-      end
-      # Calling .wait before .value! to make sure that all tasks are completed.
-      tasks_creating.map(&:wait)
-      network_interfaces = tasks_creating.map(&:value!)
-    end
-
-    def _delete_possible_network_interfaces(resource_group_name, vm_name)
-      network_interfaces = @azure_client.list_network_interfaces_by_keyword(resource_group_name, vm_name)
-      tasks = []
-      network_interfaces.each do |network_interface|
-        tasks.push(
-          Concurrent::Future.execute do
-            @azure_client.delete_network_interface(resource_group_name, network_interface[:name])
-          end
-        )
-      end
-      # Calling .wait before .wait! to make sure that all tasks are completed.
-      tasks.map(&:wait)
-      tasks.map(&:wait!)
-    end
-
-    def _get_availability_set_name(vm_props, env)
-      availability_set_name = vm_props.availability_set
-      if availability_set_name.nil?
-        unless env.nil? || env['bosh'].nil? || env['bosh']['group'].nil?
-          availability_set_name = env['bosh']['group']
-
-          # https://github.com/cloudfoundry-incubator/bosh-azure-cpi-release/issues/209
-          # On Azure the length of the availability set name must be between 1 and 80 characters.
-          # The group name which is generated by BOSH director may be too long.
-          # CPI will truncate the name to below format if the length is greater than 80.
-          # az-MD5-[LAST-40-CHARACTERS-OF-GROUP]
-          if availability_set_name.size > 80
-            md = Digest::MD5.hexdigest(availability_set_name)
-            availability_set_name = "az-#{md}-#{availability_set_name[-40..-1]}"
-          end
-        end
-      end
-
-      availability_set_name
-    end
-
     def _get_tags(vm_props)
       tags = AZURE_TAGS.dup
       custom_tags = vm_props.tags
       tags.merge!(custom_tags)
-    end
-
-    def _get_or_create_availability_set(resource_group_name, availability_set_name, vm_props, location)
-      return nil if availability_set_name.nil?
-
-      availability_set_params = {
-        name: availability_set_name,
-        location: location,
-        tags: AZURE_TAGS,
-        platform_update_domain_count: vm_props.platform_update_domain_count || default_update_domain_count,
-        platform_fault_domain_count: vm_props.platform_fault_domain_count || default_fault_domain_count,
-        managed: @use_managed_disks
-      }
-      availability_set = nil
-      flock("#{CPI_LOCK_PREFIX_AVAILABILITY_SET}-#{availability_set_name}", File::LOCK_EX) do
-        availability_set = @azure_client.get_availability_set_by_name(resource_group_name, availability_set_name)
-        if availability_set.nil?
-          @logger.info("create_availability_set - the availability set '#{availability_set_name}' doesn't exist. Will create a new one.")
-          @azure_client.create_availability_set(resource_group_name, availability_set_params)
-          availability_set = @azure_client.get_availability_set_by_name(resource_group_name, availability_set_name)
-        # In some regions, the location of availability set is case-sensitive, e.g. CanadaCentral instead of canadacentral.
-        elsif !availability_set[:location].casecmp(availability_set_params[:location]).zero?
-          cloud_error("create_availability_set - the availability set '#{availability_set_name}' already exists, but in a different location '#{availability_set[:location].downcase}' instead of '#{availability_set_params[:location].downcase}'. Please delete the availability set or choose another location.")
-        elsif !@use_managed_disks && availability_set[:managed]
-          cloud_error("create_availability_set - the availability set '#{availability_set_name}' already exists. It's not allowed to update it from managed to unmanaged.")
-        elsif @use_managed_disks && !availability_set[:managed]
-          @logger.info("create_availability_set - the availability set '#{availability_set_name}' exists, but it needs to be updated from unmanaged to managed.")
-          availability_set_params.merge!(
-            platform_update_domain_count: availability_set[:platform_update_domain_count],
-            platform_fault_domain_count: availability_set[:platform_fault_domain_count],
-            managed: true
-          )
-          @azure_client.create_availability_set(resource_group_name, availability_set_params)
-          availability_set = @azure_client.get_availability_set_by_name(resource_group_name, availability_set_name)
-        else
-          @logger.info("create_availability_set - the availability set '#{availability_set_name}' exists. No need to update.")
-        end
-      end
-      cloud_error("get_or_create_availability_set - availability set '#{availability_set_name}' is not created.") if availability_set.nil?
-      availability_set
-    end
-
-    def _delete_empty_availability_set(resource_group_name, availability_set_name)
-      flock("#{CPI_LOCK_PREFIX_AVAILABILITY_SET}-#{availability_set_name}", File::LOCK_EX | File::LOCK_NB) do
-        availability_set = @azure_client.get_availability_set_by_name(resource_group_name, availability_set_name)
-        @azure_client.delete_availability_set(resource_group_name, availability_set_name) if availability_set && availability_set[:virtual_machines].empty?
-      end
     end
 
     def _create_virtual_machine(instance_id, vm_params, network_interfaces, availability_set)
